requirements.txt
# Core Agent Framework
strands-agents>=0.1.0
strands-agents-tools>=0.1.0
bedrock-agentcore>=0.1.0

# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Data Handling
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
hypothesis>=6.100.0

# Web Interface
streamlit>=1.32.0

# AWS SDK
boto3>=1.34.0
